+++ Controller/SecurityController.php
@@ -61,9 +61,7 @@ class SecurityController extends Controller
      */
     protected function renderLogin(array $data)
     {
-        $template = sprintf('FOSUserBundle:Security:login.html.%s', $this->container->getParameter('fos_user.template.engine'));
-
-        return $this->container->get('templating')->renderResponse($template, $data);
+        return $this->container->get('templating')->renderResponse('FOSUserBundle:Security:login.html.twig', $data);
     }
 
     public function checkAction()